refactor(proxy): extract page-route and api-route handlers from src/proxy.ts
src/proxy.ts had grown to 340 lines as the single home for all proxy
orchestration: dispatch by path prefix, page-route i18n + auth + access
restriction + passkey enforcement, API-route classification + CSRF +
CORS + Bearer-bypass + session check. Extract two cohesive sub-handlers
so the orchestrator is small enough to read top-to-bottom.

- src/lib/proxy/page-route.ts (new): handlePageRoute + intl middleware,
  passkey-exempt path checks, passkey grace period, the passkey audit
  dedup map and recordPasskeyAuditEmit, clearAuthSessionCookies. The
  ProxyOptions type lives here as the primary consumer.
- src/lib/proxy/api-route.ts (new): handleApiAuth, with no behavior
  change. Imports cors-gate, auth-gate, csrf-gate, route-policy, etc.
  exactly as before.
- src/proxy.ts: thin dispatcher (40 lines) — API_ROOT prefix → api-route
  handler, otherwise → page-route handler. Test shim re-exports preserved
  so src/__tests__/proxy.test.ts continues to import via this path.

E F1 from PR #398 review log. ≤100-line goal hit (340 → 40 in proxy.ts).
No behavior change; full proxy.test.ts and consent route.test.ts pass.
